gestion des utilasateur

+++ categorie.php
@@ -1,6 +1,7 @@
 <?php
 include "db.php";
-
+session_start();
+$role=$_SESSION['role'];
 if (isset($_GET['delete'])) {
     $id = intval($_GET['delete']);
     try {
@@ -96,15 +97,17 @@ try {
                             <i class="fa-solid fa-file-lines"></i> Articles
                         </a>
                     </li>
+                    <?php if (isset($_SESSION['role']) && $_SESSION['role'] == "author"): ?>
                     <li>
                         <a href="comments.php" class="flex items-center gap-3 p-3 hover:text-white transition-colors pl-8">
                             <i class="fa-regular fa-comments"></i> Comments
                         </a>
                     </li>
+                    <?php  endif;  ?>
                 </ul>
                  <div class="mt-4 px-6 mb-2 text-xs uppercase text-gray-500 font-semibold">Modules</div>
             <ul class="px-3">
-                <li><a href="#" class="flex items-center gap-3 p-3 hover:text-white rounded-lg"><i class="fa-solid fa-users"></i> Users</a></li>
+                <li><a href="users.php" class="flex items-center gap-3 p-3 hover:text-white rounded-lg"><i class="fa-solid fa-users"></i> Users</a></li>
             </ul>
         </nav>
         
@@ -112,7 +115,7 @@ try {
             <div class="flex items-center gap-3">
                 <img src="https://ui-avatars.com/api/?name=Admin&background=random" class="w-10 h-10 rounded-full bg-blue-100">
                 <div class="overflow-hidden">
-                    <h4 class="text-sm font-white text-white">Admin</h4>
+                    <h4 class="text-sm font-white text-white"><?= $role ?></h4>
                 </div>
                <a href="login.php">
                 <button class="ml-auto text-gray-500 hover:text-red-400"><i class="fa-solid fa-power-off"></i></button>
@@ -159,7 +162,9 @@ try {
                             </th>
                             <th class="p-4">Category Name</th>
                             <th class="p-4">Description</th>
+                            <?php if (isset($_SESSION['role']) && $_SESSION['role'] == "author"): ?>
                             <th class="p-4 text-right">Actions</th>
+                            <?php endif; ?>
                         </tr>
                     </thead>
                     <tbody class="divide-y divide-gray-100 text-sm text-gray-700">
@@ -187,17 +192,19 @@ try {
                                 </td>
 
                                 <td class="p-4 text-right">
-                                    <div class="flex items-center justify-end gap-2">
-                                        <a href="editCategory.php?id=<?= $cat['id_cat'] ?>" class="w-10 h-8 rounded bg-purple-600 text-white hover:bg-purple-700 flex items-center justify-center shadow-sm">
-                                            <i class="fa-solid fa-pencil"></i>
-                                        </a>
+                                    <?php if (isset($_SESSION['role']) && $_SESSION['role'] == "admin"): ?>
+                                        <div class="flex items-center justify-end gap-2">
+                                            <a href="editCategory.php?id=<?= $cat['id_cat'] ?>" class="w-10 h-8 rounded bg-purple-600 text-white hover:bg-purple-700 flex items-center justify-center shadow-sm">
+                                                <i class="fa-solid fa-pencil"></i>
+                                            </a>
 
-                                        <a href="categorie.php?delete=<?= $cat['id_cat'] ?>" 
-                                        onclick="return confirm('Are you sure you want to delete this category?')"
-                                        class="inline-flex w-8 h-8 rounded bg-pink-500 text-white hover:bg-pink-600 items-center justify-center shadow-sm">
-                                            <i class="fa-regular fa-trash-can"></i>
-                                        </a>
-                                    </div>
+                                            <a href="categorie.php?delete=<?= $cat['id_cat'] ?>" 
+                                            onclick="return confirm('Are you sure you want to delete this category?')"
+                                            class="inline-flex w-8 h-8 rounded bg-pink-500 text-white hover:bg-pink-600 items-center justify-center shadow-sm">
+                                                <i class="fa-regular fa-trash-can"></i>
+                                            </a>
+                                        </div>
+                                        <?php endif; ?>
                              </td>
                             </tr>
                             <?php endforeach; ?>
